+++ packages/client/auth-core/src/index.ts
@@ -1,2 +1,2 @@
 export * from "./services";
-export * from "./utils";
+export * from "./types";
