+++ packages/client/auth-core/src/types/user.ts
@@ -1,0 +1,16 @@
+export type FarcasterMetadata = {
+    fid: string;
+    username: string;
+    bio: string;
+    displayName: string;
+    pfpUrl: string;
+    custody: string;
+    verifications: string[];
+};
+
+export type SDKExternalUser = {
+    id: string;
+    email?: string;
+    phoneNumber?: string;
+    farcaster?: FarcasterMetadata;
+};
