+++ packages/common/base/src/types/Crossmint.ts
@@ -1,22 +1,5 @@
 import { type ValidateAPIKeyPrefixExpectations, validateAPIKey } from "@/apiKey";
 
-export type FarcasterMetadata = {
-    fid: string;
-    username: string;
-    bio: string;
-    displayName: string;
-    pfpUrl: string;
-    custody: string;
-    verifications: string[];
-};
-
-export type SDKExternalUser = {
-    id: string;
-    email?: string;
-    phoneNumber?: string;
-    farcaster?: FarcasterMetadata;
-};
-
 export type Crossmint = {
     apiKey: string;
     jwt?: string;
